:sparkles: Refactor auth forms to unify validation and data types
- Replaced `LoginFormType` and `RegisterFormType`
with a unified `AuthFormType` for consistency across authentication forms.

- Updated `LoginForm` and `RegisterForm` components
to use `AuthFormType` instead of individual form types.

- Consolidated validation logic by deleting `ValidationLogin.ts` and `ValidationRegister.ts`,
replacing them with a shared `ValidationAuth.ts` for reusable validation.

- Adjusted `login` function to use `email` instead of `username`.
